rename general document and layout tests
